Move the DeepSeek-V4 compressor tests out of CPU CI, which cannot import or run the compressor
